# services/web-app/openapi.yaml
openapi: 3.0.3
info:
  title: RespiraAlly API
  description: API for the RespiraAlly respiratory care management system
  version: 1.0.0
  contact:
    name: RespiraAlly Team
    email: support@respiraally.com
servers:
  - url: http://localhost:5000/api/v1
    description: Development server
  - url: https://api.respiraally.com/api/v1
    description: Production server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              example: "INVALID_INPUT"
            message:
              type: string
              example: "Request body cannot be empty"
...
    CATQuestionnaire:
      type: object
      properties:
        record_id:
          type: integer
          example: 1
        record_date:
          type: string
          format: date
          example: "2025-01-15"
        total_score:
          type: integer
          example: 15
        scores:
          type: object
          properties:
            cough:
              type: integer
              minimum: 0
              maximum: 5
              example: 2
            phlegm:
              type: integer
              minimum: 0
              maximum: 5
              example: 1
            chest:
              type: integer
              minimum: 0
              maximum: 5
              example: 3
            breath:
              type: integer
              minimum: 0
              maximum: 5
              example: 2
            limit:
              type: integer
              minimum: 0
              maximum: 5
              example: 1
            confidence:
              type: integer
              minimum: 0
              maximum: 5
              example: 2
            sleep:
              type: integer
              minimum: 0
              maximum: 5
              example: 2
            energy:
              type: integer
              minimum: 0
              maximum: 5
              example: 2
...
    PaginationMeta:
      type: object
      properties:
        total_items:
          type: integer
          example: 100
        total_pages:
          type: integer
          example: 10
        current_page:
          type: integer
          example: 1
        per_page:
          type: integer
          example: 10
paths:
...
  /patients/{patient_id}/questionnaires/cat:
    get:
      summary: "獲取 CAT 問卷歷史記錄"
      description: "獲取指定病患的所有歷史 CAT 問卷記錄，用於繪製分數趨勢圖"
      tags: ["Health Data & Questionnaires"]
      security:
        - bearerAuth: []
      parameters:
        - name: patient_id
          in: path
          required: true
          schema:
            type: integer
          description: "病患的 user_id"
        - name: page
          in: query
          schema:
            type: integer
            default: 1
          description: "頁碼"
        - name: per_page
          in: query
          schema:
            type: integer
            default: 12
          description: "每頁數量"
      responses:
        '200':
          description: "成功獲取 CAT 問卷歷史"
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/CATQuestionnaire'
                  pagination:
                    $ref: '#/components/schemas/PaginationMeta'
        '401':
          description: "Token 無效或未提供"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: "沒有權限查看此病患的記錄"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: "找不到該病患"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    post:
      summary: "提交 CAT 問卷"
      description: "提交一份新的 CAT 問卷。如果當月已存在記錄，將回傳 409 Conflict"
      tags: ["Health Data & Questionnaires"]
      security:
        - bearerAuth: []
      parameters:
        - name: patient_id
          in: path
          required: true
          schema:
            type: integer
          description: "病患的 user_id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [record_date, cough_score, phlegm_score, chest_score, breath_score, limit_score, confidence_score, sleep_score, energy_score]
              properties:
                record_date:
                  type: string
                  format: date
                  description: "記錄日期 (YYYY-MM-DD)"
                cough_score:
                  type: integer
                  minimum: 0
                  maximum: 5
                  description: "咳嗽程度 (0-5)"
                phlegm_score:
                  type: integer
                  minimum: 0
                  maximum: 5
                  description: "痰液狀況 (0-5)"
                chest_score:
                  type: integer
                  minimum: 0
                  maximum: 5
                  description: "胸口緊繃 (0-5)"
                breath_score:
                  type: integer
                  minimum: 0
                  maximum: 5
                  description: "氣喘程度 (0-5)"
                limit_score:
                  type: integer
                  minimum: 0
                  maximum: 5
                  description: "日常活動限制 (0-5)"
                confidence_score:
                  type: integer
                  minimum: 0
                  maximum: 5
                  description: "信心外出 (0-5)"
                sleep_score:
                  type: integer
                  minimum: 0
                  maximum: 5
                  description: "睡眠品質 (0-5)"
                energy_score:
                  type: integer
                  minimum: 0
                  maximum: 5
                  description: "精力程度 (0-5)"
      responses:
        '201':
          description: "CAT 問卷提交成功"
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      record_id:
                        type: integer
                      total_score:
                        type: integer
                      message:
                        type: string
        '400':
          description: "請求 Body 的資料格式或分數範圍錯誤"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: "Token 無效或未提供"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: "試圖為其他病患提交問卷"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: "找不到具有指定 patient_id 的病患"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: "指定月份已存在問卷記錄"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /patients/{patient_id}/questionnaires/cat/{year}/{month}:
    put:
      summary: "更新指定月份的 CAT 問卷"
      description: "更新指定年月份的 CAT 問卷"
      tags: ["Health Data & Questionnaires"]
      security:
        - bearerAuth: []
      parameters:
        - name: patient_id
          in: path
          required: true
          schema:
            type: integer
          description: "病患的 user_id"
        - name: year
          in: path
          required: true
          schema:
            type: integer
          description: "要更新的問卷年份 (e.g., 2025)"
        - name: month
          in: path
          required: true
          schema:
            type: integer
            minimum: 1
            maximum: 12
          description: "要更新的問卷月份 (1-12)"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [record_date, cough_score, phlegm_score, chest_score, breath_score, limit_score, confidence_score, sleep_score, energy_score]
              properties:
                record_date:
                  type: string
                  format: date
                phlegm_score:
                  type: integer
                  minimum: 0
                  maximum: 5
                chest_score:
                  type: integer
                  minimum: 0
                  maximum: 5
                breath_score:
                  type: integer
                  minimum: 0
                  maximum: 5
                limit_score:
                  type: integer
                  minimum: 0
                  maximum: 5
                confidence_score:
                  type: integer
                  minimum: 0
                  maximum: 5
                sleep_score:
                  type: integer
                  minimum: 0
                  maximum: 5
                energy_score:
                  type: integer
                  minimum: 0
                  maximum: 5
      responses:
        '200':
          description: "CAT 問卷更新成功"
        '400':
          description: "請求 Body 的資料格式或日期格式錯誤"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: "Token 無效或未提供"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: "試圖為其他病患更新問卷"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: "找不到指定月份的問卷記錄"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # MMRC Questionnaire endpoints
...